# docs/openapi.yaml
openapi: 3.1.0
info:
  title: TechNest Shop API
  version: 1.0.0
  description: >
    Versioned OpenAPI contract for the current TechNest Shop backend. This file is
    the authoritative machine-readable API contract; README keeps the operational
    snapshot and docs/API_CONTRACT.md keeps a concise human summary.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
...
paths:
...
  /api/products:
    get:
      tags: [Catalog]
      operationId: listProducts
      summary: List products
      security: []
      parameters:
        - in: query
          name: categoryId
          schema:
            type: integer
            format: int64
            minimum: 1
        - in: query
          name: cat
          schema:
            type: string
          description: Category name alias, for example `laptop`, `phone`, `all`.
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: minPrice
          schema:
            type: string
          description: Decimal string.
        - in: query
          name: maxPrice
          schema:
            type: string
          description: Decimal string.
        - in: query
          name: brand
          schema:
            type: string
      responses:
        '200':
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    post:
      tags: [Catalog]
      operationId: createProduct
      summary: Create product
      description: Staff or admin only.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpsertRequest'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/products/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Catalog]
      operationId: getProduct
      summary: Get product by id
      security: []
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Catalog]
      operationId: updateProduct
      summary: Update product
      description: Staff or admin only.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpsertRequest'
      responses:
        '200':
          description: Product updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Catalog]
      operationId: deleteProduct
      summary: Delete product
      description: Admin only.
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Product deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/reviews/product/{productId}:
    parameters:
      - $ref: '#/components/parameters/ProductIdPath'
    get:
      tags: [Reviews]
      operationId: listProductReviews
      summary: List approved reviews for a product
      security: []
      responses:
        '200':
          description: Approved reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReviewResponse'
    post:
      tags: [Reviews]
      operationId: createReview
      summary: Create review for a product
      description: >
        Authenticated user only. Business logic only allows customers to review
        products from delivered and paid orders. New reviews start as pending.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      responses:
        '200':
          description: Review submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewMutationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /api/reviews/manage/product/{productId}:
    parameters:
      - $ref: '#/components/parameters/ProductIdPath'
    get:
      tags: [Reviews]
      operationId: listManagementReviews
      summary: List all reviews for moderation
      description: Staff or admin only.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Approved and pending reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReviewResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    ProductIdPath:
      in: path
      name: productId
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
  responses:
    BadRequest:
      description: Request validation failed or payload is invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Unauthorized:
      description: Authentication is required or credentials are invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Forbidden:
      description: Authenticated but not allowed to perform this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Conflict:
      description: Request conflicts with existing data or business rules
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
  schemas:
    ApiErrorResponse:
      type: object
      properties:
        message:
          type: string
        status:
          type: integer
        error:
          type: string
        validationErrors:
          type: object
          additionalProperties:
            type: string
      required: [message, status, error]
...
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          maxLength: 180
        price:
          type: number
        imageUrl:
          type: string
          maxLength: 255
        categoryId:
          type: integer
          format: int64
        categoryName:
          type: string
        quantity:
          type: integer
        descriptionShort:
          type: string
        descriptionLong:
          type: string
      required: [id, name, price, quantity]
    ProductUpsertRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 180
        price:
          type: number
          minimum: 0
        imageUrl:
          type: string
          maxLength: 255
        categoryId:
          type: integer
          format: int64
          minimum: 1
        quantity:
          type: integer
          minimum: 0
        descriptionShort:
          type: string
          maxLength: 2000
        descriptionLong:
          type: string
          maxLength: 10000
      required: [name, price, quantity]
...
    ReviewReplyResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        staffId:
          type: integer
          format: int64
        body:
          type: string
      required: [id, staffId, body]
    ReviewResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        userName:
          type: string
        userAvatar:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        title:
          type: string
        body:
          type: string
        isApproved:
          type: boolean
        createdAt:
          type: string
        reply:
          oneOf:
            - $ref: '#/components/schemas/ReviewReplyResponse'
            - type: 'null'
      required:
        - id
        - userId
        - userName
        - userAvatar
        - rating
        - title
        - body
        - isApproved
        - createdAt
        - reply
...
    CreateReviewRequest:
      type: object
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        title:
          type: string
          maxLength: 150
        body:
          type: string
          maxLength: 2000
      required: [rating, body]
...
    ReviewMutationResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        message:
          type: string
      required: [id, message]
...